Checkpoint #6
Implement SP update
