+++ components/security-mgt/org.wso2.carbon.security.mgt/src/main/java/org/wso2/carbon/security/internal/SecurityMgtServiceComponent.java
@@ -36,6 +36,8 @@ import org.wso2.carbon.registry.core.service.RegistryService;
 import org.wso2.carbon.registry.core.service.TenantRegistryLoader;
 import org.wso2.carbon.security.SecurityServiceHolder;
 import org.wso2.carbon.security.config.SecurityConfigAdmin;
+import org.wso2.carbon.security.sts.service.STSAdminServiceImpl;
+import org.wso2.carbon.security.sts.service.STSAdminServiceInterface;
 import org.wso2.carbon.user.core.service.RealmService;
 import org.wso2.carbon.utils.Axis2ConfigurationContextObserver;
 import org.wso2.carbon.utils.ConfigurationContextService;
@@ -73,6 +75,7 @@ public class SecurityMgtServiceComponent {
                     new SecurityConfigAdmin(mainAxisConfig, registryService.getConfigSystemRegistry(), null), null);
             bundleCtx.registerService(Axis2ConfigurationContextObserver.class.getName(),
                     new SecurityAxis2ConfigurationContextObserver(), null);
+            bundleCtx.registerService(STSAdminServiceInterface.class, new STSAdminServiceImpl(), null);
             log.debug("Security Mgt bundle is activated");
         } catch (Throwable e) {
             log.error("Failed to activate SecurityMgtServiceComponent", e);
